+++ node/src/triple.rs
@@ -36,11 +36,15 @@ pub async fn run_triple_generation(
 /// There is, however, a chance that the same participant generates an ID that
 /// already existed before, so the existence of a triple of such an ID must be
 /// checked before using it.
+///
+/// TODO(#10): is this a good way to generate IDs?
+#[cfg(test)]
 pub fn generate_triple_id(me: ParticipantId) -> u64 {
     (rand::random::<u64>() >> 12) | ((me.0 as u64) << 52)
 }
 
-/// Generates a cait-sith triple.
+/// Generates many cait-sith triples at once. This can significantly save the
+/// *number* of network messages.
 pub async fn run_many_triple_generation<const N: usize>(
     channel: NetworkTaskChannel,
     participants: Vec<ParticipantId>,
@@ -62,6 +66,10 @@ pub async fn run_many_triple_generation<const N: usize>(
     Ok(triples)
 }
 
+/// A very simple in-memory triple store. Stores our own triples separately from
+/// others. For our own triples, it is a queue that allows async waiting. For
+/// triples generated by others, we store them so that when they perform
+/// presignatures and give us the triple IDs we would have them.
 pub struct SimpleTripleStore {
     my_triples_receiver: flume::Receiver<(u64, TripleGenerationOutput<Secp256k1>)>,
     my_triples_sender: flume::Sender<(u64, TripleGenerationOutput<Secp256k1>)>,
@@ -78,10 +86,12 @@ impl SimpleTripleStore {
         }
     }
 
+    /// An approximate count of triples we generated.
     pub fn my_triples_count(&self) -> usize {
         self.my_triples_receiver.len()
     }
 
+    /// Removes their triple by ID, erroring if the ID doesn't exist.
     pub fn take_their_triple(&self, id: u64) -> anyhow::Result<TripleGenerationOutput<Secp256k1>> {
         self.their_triples
             .lock()
@@ -90,20 +100,25 @@ impl SimpleTripleStore {
             .ok_or_else(|| anyhow::anyhow!("Triple with ID {} not found in the triple store", id))
     }
 
+    /// Removes a triple generated by us, returning the ID and triple.
+    /// If a triple is not yet available (we've exhausted all of them),
+    /// this asynchronously blocks until one is available. An error result
+    /// is only emitted during shutdown.
     pub async fn take_my_triple(&self) -> anyhow::Result<(u64, TripleGenerationOutput<Secp256k1>)> {
-        Ok(self
-            .my_triples_receiver
+        self.my_triples_receiver
             .recv_async()
             .await
-            .context("Receiving triple from triple store")?)
+            .context("Receiving triple from triple store")
     }
 
+    /// Add a triple we generated.
     pub fn add_my_triple(&self, id: u64, triple: TripleGenerationOutput<Secp256k1>) {
         self.my_triples_sender
             .send((id, triple))
             .expect("Sending triple to triple store");
     }
 
+    /// Add a triple that we have helped others generate.
     pub fn add_their_triple(&self, id: u64, triple: TripleGenerationOutput<Secp256k1>) {
         self.their_triples.lock().unwrap().insert(id, triple);
     }
@@ -111,6 +126,9 @@ impl SimpleTripleStore {
 
 pub const SUPPORTED_TRIPLE_GENERATION_BATCH_SIZE: usize = 64;
 
+/// Continuously runs triple generation in the background, using the number of threads
+/// specified in the config, trying to maintain some number of available triples all the
+/// time as specified in the config.
 pub async fn run_background_triple_generation(
     client: Arc<MeshNetworkClient>,
     threshold: usize,
@@ -119,6 +137,7 @@ pub async fn run_background_triple_generation(
 ) -> anyhow::Result<()> {
     // Start with unix epoch millis timestamp as the ID, as that is unlikely to
     // collide with what we've used so far.
+    // TODO(#10): is this a good way to generate IDs?
     let mut next_id_to_use = std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .expect("SystemTime before UNIX EPOCH")
@@ -131,6 +150,10 @@ pub async fn run_background_triple_generation(
         metrics::MPC_OWNED_NUM_TRIPLES_AVAILABLE.set(my_triples_count as i64);
         if my_triples_count + in_flight_generations.num_in_flight()
             < config.desired_triples_to_buffer
+            // There's no point to issue way too many in-flight computations, as they
+            // will just be limited by the concurrency anyway.
+            && in_flight_generations.num_in_flight()
+                < config.concurrency * 2 * SUPPORTED_TRIPLE_GENERATION_BATCH_SIZE
         {
             let id_start = next_id_to_use;
             let id_end = next_id_to_use + SUPPORTED_TRIPLE_GENERATION_BATCH_SIZE as u64;
@@ -161,11 +184,12 @@ pub async fn run_background_triple_generation(
                 anyhow::Ok(())
             });
         } else {
-            tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+            tokio::time::sleep(std::time::Duration::from_millis(100)).await;
         }
     }
 }
 
+/// Tracks number of in-flight generations so we don't generate too many at the same time.
 struct InFlightGenerationTracker {
     generations_in_flight: Arc<AtomicUsize>,
 }
@@ -187,6 +211,7 @@ impl InFlightGenerationTracker {
     }
 }
 
+/// Drop guard to increment and decrement number of generations in flight.
 struct InFlightGenerations {
     generations_in_flight: Arc<AtomicUsize>,
     count: usize,
