+++ node/src/web.rs
@@ -1,10 +1,12 @@
 use crate::config::WebUIConfig;
-use crate::hash::sha256hash;
 use crate::mpc_client::MpcClient;
 use crate::tracking::{self, TaskHandle};
 use actix_web::error::ErrorInternalServerError;
 use actix_web::{get, web};
 use futures::{stream, StreamExt, TryStreamExt};
+use k256::elliptic_curve::scalar::FromUintUnchecked;
+use k256::sha2::{Digest, Sha256};
+use k256::{Scalar, U256};
 use prometheus::{default_registry, Encoder};
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
@@ -56,7 +58,16 @@ async fn debug_sign(
             ))
         })
         .await
-        .map_err(|e| ErrorInternalServerError(e))
+        .map_err(ErrorInternalServerError)
+}
+
+fn sha256hash(data: &[u8]) -> k256::Scalar {
+    let mut hasher = Sha256::new();
+    hasher.update(data);
+    let result = hasher.finalize();
+    let mut bytes = [0u8; 32];
+    bytes.copy_from_slice(&result);
+    Scalar::from_uint_unchecked(U256::from_be_slice(&bytes))
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
